move models under db/
