+++ templates/nextjs-api-routes/components/EachToDo.tsx
@@ -1,6 +1,6 @@
 import Image from "next/image";
 import React from "react";
-import { TodoItem } from "../models/todoItems";
+import { TodoItem } from "../db/models/todoItems";
 import styles from "../styles/EachToDo.module.css";
 
 type Props = {
